Validation, user folders, logging out
Validation to check if the file already exists in the users directory.

Users, when creating an account, will now have their own unique folder, to hold their files.

+++ PHP/Scripts/fileUploadScript.php
@@ -1,11 +1,39 @@
 <?php
+    include_once '../Classes/SessionsClass.php';
+
+    $sessionsStart = new sessionClass();
+    $sessionsStart -> startSession();
+    
     $numOfFiles = count($_FILES["filesToUpload"]["name"]);
 
-    for ($i = 0; $i < $numOfFiles; $i++){
-        $targetDir = "../../uploads/";
-        $targetFile = $targetDir . basename($_FILES["filesToUpload"]["name"][$i]);
-        if(move_uploaded_file($_FILES["filesToUpload"]["tmp_name"][$i], $targetFile)){
-            echo "You did it Gary. Check your XAMPP folder now.";
+    $supportedFormats = array(
+        "PNG"=>"png",
+        "JPEG"=>"jpeg",
+        "GIF"=>"gif",
+        "JPG"=>"jpg"
+    );
+    if ($numOfFiles == 1 && $_FILES["filesToUpload"]["name"][0] == null){
+        echo "No Files Selected.:";
+    } 
+    else{
+        for ($i = 0; $i < $numOfFiles; $i++){
+            $targetDir = $_SESSION['UserAccount']['UploadDirectory']."/";
+            $targetFile = $targetDir . basename($_FILES["filesToUpload"]["name"][$i]);
+            $imageFileType = strtolower(pathinfo($targetFile,PATHINFO_EXTENSION));
+            if(in_array($imageFileType, $supportedFormats, true)){
+                if(!file_exists($targetFile)){
+                    if(move_uploaded_file($_FILES["filesToUpload"]["tmp_name"][$i], $targetFile)){
+                        $messageSuccess = "File: " . basename($_FILES["filesToUpload"]["name"][$i]) . " uploaded successfully <br>";
+                        echo $messageSuccess;
+                    } else{
+                        echo "Something went wrong.";
+                    }
+                } else{
+                    echo "This file already exists";
+                }
+            } else{
+                echo "This file <strong>" . basename($_FILES["filesToUpload"]["name"][$i]) . "</strong> is not currently supported at this stage. We are very sorry";
+            }
         }
     }
 ?>
